requirements.txt
# Discord dependencies
discord
discord.py

# Additional dependencies
python-dotenv
requests